improve component filing hierarchy
