+++ src/components/bulletin-components/Bulletin.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import BulletinHeader from './BulletinHeader';
 import BulletinMessage from './BulletinMessage';
-import AvatarIcon from './AvatarIcon';
+import AvatarIcon from '../AvatarIcon';
+import { Link } from 'react-router-dom';
 
 type BulletinProps = {
   header: string;
@@ -16,9 +17,21 @@ export default class Bulletin extends React.Component<BulletinProps> {
     return (
       <div>
         <tr>
-          <td>
-            <BulletinHeader header={header} id={id} />
-          </td>
+          <Link
+            to={{
+              pathname: `/comment/${id}`,
+              state: {
+                header,
+                id,
+                ownerId,
+                message
+              }
+            }}
+          >
+            <td>
+              <BulletinHeader header={header} />
+            </td>
+          </Link>
           <td>
             <AvatarIcon id={ownerId} />
           </td>
