+++ src/components/bulletin-components/BulletinHeader.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+type BulletinHeaderProps = {
+  header: string;
+};
+
+export default class BulletinHeader extends React.Component<BulletinHeaderProps> {
+  public render() {
+    const { header } = this.props;
+    return <div>{header}</div>;
+  }
+}
