+++ src/components/bulletin-components/BulletinMessage.tsx
