+++ src/components/bulletin-components/BulletinSubmit.tsx
